Israel Martinez Modificacion cambipo de modelo

+++ AppGia/Models/Empresa_Proyecto.cs
@@ -11,9 +11,8 @@ namespace AppGia.Models
 
         public bool activo { get; set; }
 
-        public int empresa_id { get; set; }
+        public long empresa_id { get; set; }
 
-        public int proyecto_id { get; set; }
-
+        public long proyecto_id { get; set; }
     }
 }
